Check if a namespace is loaded

*_namespace() check if an argument has zero length.

Usage

check_namespace(x, null_ok = FALSE, .names = deparse(substitute(x)))

test_namespace(x, null_ok = FALSE, .names = deparse(substitute(x)))

assert_namespace(
  x,
  null_ok = FALSE,
  .names = deparse(substitute(x)),
  .var.name = checkmate::vname(x),
  add = NULL
)

expect_namespace(
  x,
  null_ok = FALSE,
  .names = deparse(substitute(x)),
  info = NULL,
  label = checkmate::vname(x)
)

Arguments

x

Any R object.

null_ok

(Optional) A logical flag indicating if NULL values are allowed (Default: FALSE).

.names

(Optional) A character vector specifying the names of the objects being tested. This argument is used internally and should not be set by the user.

.var.name

(Optional) A character string indicating the name of the checked object to print in assertions. This argument is used internally and should not be set by the user.

add

(Optional) A AssertCollection object indicating the collection to store assertion messages. This argument is used internally and should not be set by the user.

info

(Optional) A character string with extra information to be included in the message for the `testthat“ reporter. This argument is used internally and should not be set by the user.

label

(Optional) A character string indicating the name of the checked object to print in messages. This argument is used internally and should not be set by the user.

Examples

test_namespace("TeSt")
#> [1] FALSE
#> [1] FALSE # Expected

test_namespace("prettycheck")
#> [1] TRUE
#> [1] TRUE # Expected